+++ react-components/src/architecture/concrete/cropping/CropBoxTool.ts
@@ -5,9 +5,9 @@
 import { type BaseCommand } from '../../base/commands/BaseCommand';
 import { type BaseCreator } from '../../base/domainObjectsHelpers/BaseCreator';
 import { type TranslateKey } from '../../base/utilities/TranslateKey';
-import { PrimitiveEditTool } from '../boxAndLines/PrimitiveEditTool';
-import { type PrimitiveType } from '../boxAndLines/PrimitiveType';
-import { BoxCreator } from '../boxAndLines/BoxCreator';
+import { PrimitiveEditTool } from '../primitives/PrimitiveEditTool';
+import { type PrimitiveType } from '../primitives/PrimitiveType';
+import { BoxCreator } from '../primitives/BoxCreator';
 import { CropBoxDomainObject } from './CropBoxDomainObject';
 import { CropCommand } from './CropCommand';
 import { ShowCropBoxOnTopCommand } from './commands/ShowCropBoxOnTopCommand';
