+++ react-components/src/architecture/concrete/measurements/MeasureBoxDomainObject.ts
@@ -2,8 +2,8 @@
  * Copyright 2024 Cognite AS
  */
 
-import { PrimitiveType } from '../boxAndLines/PrimitiveType';
-import { BoxDomainObject } from '../boxAndLines/BoxDomainObject';
+import { PrimitiveType } from '../primitives/PrimitiveType';
+import { BoxDomainObject } from '../primitives/BoxDomainObject';
 import { Color } from 'three';
 import { getIconByPrimitiveType } from './getIconByPrimitiveType';
 import { PanelInfo } from '../../base/domainObjectsHelpers/PanelInfo';
