+++ react-components/src/architecture/concrete/measurements/MeasurementTool.ts
@@ -8,12 +8,12 @@ import { ShowMeasurementsOnTopCommand } from './commands/ShowMeasurementsOnTopCo
 import { SetMeasurementTypeCommand } from './commands/SetMeasurementTypeCommand';
 import { type TranslateKey } from '../../base/utilities/TranslateKey';
 import { ToggleMetricUnitsCommand } from '../../base/concreteCommands/ToggleMetricUnitsCommand';
-import { PrimitiveEditTool } from '../boxAndLines/PrimitiveEditTool';
+import { PrimitiveEditTool } from '../primitives/PrimitiveEditTool';
 import { MeasureLineDomainObject } from './MeasureLineDomainObject';
 import { MeasureBoxDomainObject } from './MeasureBoxDomainObject';
-import { PrimitiveType } from '../boxAndLines/PrimitiveType';
-import { BoxCreator } from '../boxAndLines/BoxCreator';
-import { LineCreator } from '../boxAndLines/LineCreator';
+import { PrimitiveType } from '../primitives/PrimitiveType';
+import { BoxCreator } from '../primitives/BoxCreator';
+import { LineCreator } from '../primitives/LineCreator';
 import { type VisualDomainObject } from '../../base/domainObjects/VisualDomainObject';
 
 export class MeasurementTool extends PrimitiveEditTool {
