+++ react-components/src/architecture/concrete/primitives/BoxDomainObject.ts
