+++ react-components/src/architecture/concrete/primitives/BoxRenderStyle.ts
@@ -3,10 +3,10 @@
  */
 
 import { cloneDeep } from 'lodash';
-import { TextRenderStyle } from './TextRenderStyle';
+import { PrimitiveRenderStyle } from './PrimitiveRenderStyle';
 import { type RenderStyle } from '../../base/domainObjectsHelpers/RenderStyle';
 
-export class BoxRenderStyle extends TextRenderStyle {
+export class BoxRenderStyle extends PrimitiveRenderStyle {
   // ==================================================
   // INSTANCE FIELDS
   // ==================================================
