+++ react-components/src/architecture/concrete/primitives/BoxView.ts
