+++ react-components/src/architecture/concrete/primitives/LineDomainObject.ts
