+++ react-components/src/architecture/concrete/primitives/LineView.ts
